requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
chromadb
pypdf
sentence-transformers
firebase-admin
